[mapleall]Internal commit msg:     [ICSL] Clean Code 27     cg_feature_tlswarmupopt3     [mapleall] cleancode2  31     [mplcg] fix DTS2023041501584     [codecheck me_slp]     cleancode     DTS2023041212357     [mplcg] update label2bb map in ico     [mplcg] fix extendShiftOptPattern bug in globalopt     [cg] framefinalize sub need split     [ICSL] Clean Code 26     [mplcg] Ubfx merged into extendshiftoperand in prop     [mapleall] cleancode2  30     [pgo] only save last samples of pgo gen     [ICSL] Clean Code 23     [mplcg] fno-semantic-interposition use single .set     [mplcg] CGIR Verify: add splitFunc     [ICSL] Clean Code 22     [ICSL] Clean Code 25     [cg]Code alignment parameter tuning     fix constant fold sub bug     fPIE_in_remat     [ICSL] Clean Code 24     [codecheck const me_rename2preg.cpp me_merge_stmts.cpp]     [mapleall] enhace ccmp in ico     [mapleall] enhance bior and shl in irmap     [mapleall] bugfix update ssa in optimizeCFG     cg_bugfix_support_noplt_call_after_ra     fix hdse invoke bug     [mapleall] cleancode2  29     [driver]fixbug use release in debug or use debug in release     [mplcg] add HasBody check for enable -fpie option     [cg] ra code check 0411     [codecheck me_slp]     [add as_const]     [be] The becommon::GetTypeSize interface is migrated to the MIRType::GetSize interface.     [driver]support ake .o and real .o     [driver]dts-option     [cg] modify regsaves     [mapleall] cleancode2  26     [add const]     cg_bugfix_DTS2023040615529     [mplcg] Enhance ExtendShiftPattern: ubfx merged into bitshiftoperand     cg_bugfix_DTS2023040614902     [mplcg] optimize conflict graph construction     [mapleall] cleancode2  27     [mapleall] cleancode2  25     [ICSL] Clean Code 20     [mplcg] fix -fno-semantic-interposition     [mplcg] enable nosemanticInterposition on lto     [cg]add lto test case     [ptr2ref]     [add const, ptr2ref]     [mapleall] bugfix objsize     [driver]only enbale -g in O0     [mplcg] fix bug in patterns of cfgo and globalopt     enhance me sink to get better code     [mplcg] fix bug in peephole LslAndToUbfizPattern     refactor of validbit opt     [ICSL] Clean Code 21     [ICSL] add UTs for -fno-semantic-interposition     [driver]remove whitespace     [driver]only enbale -g in O0     [codecheck 20]     [cleancode m21]     [ICSL] Clean Code 19     [mapleall] cleancode2  24     [stack protect] retVar does not need to set stack protect     [cg] codecheck fix     [ra] large func, spill mem opnd bugfix     [mapleall] IntVal: fix memory leaks and buffer overflow     [WIP] [mapleall] int128: support int128 in FE     [codecheck c0]     [mplcg] sps cfg fix && pgo profile update     [driver]fixbug in option sort     [ICSL] Clean Code 18     [mplcg] fix fno-semantic-interposition issues:     [ICSL] Clean Code 12-1 fix G.EXP.04 warning     fix the offset calculation of the var_arg stack     [cg] targetprop need prop vary type     [mplcg]litepgo bb spilt bugfix     [ICSL] Clean Code 17     [ICSL] Clean Code 16     [be] The becommon::GetTypeAligninterface is migrated to the MIRType::GetAlign interface.     [codecheck a]     [ICSL] Clean Code 15     [mapleall] CGIR Verify Info Dump     [driver]--help     [ddaa] back prop nads when reaching global node     [cg] param fix large mem opnd     clean code     [mapleall] cleancode2  23     DTS2023032815634     [codecheck 18]     [mapleall] bugfix opt redundant branch with zext opnd     [mapleall] bugfix ico     [codecheck 19]     [alias] set next level NADS  for peg node assigned to global     [mplcg] fix initial-exec tls model insn issue     [mplcg] fix bug that the commonExitBB infos are not properly maintained after cfgo     [me] block merge between extensions with different sign     [driver]fix bug in driver (-fo-pic,maple *.c > 1.txt, ldoption)

+++ src/mapleall/maple_ir/src/lexer.cpp
@@ -247,7 +247,7 @@ TokenKind MIRLexer::GetLongHexConst(uint32 valStart, bool negative) {
 }
 
 TokenKind MIRLexer::GetIntConst(uint32 valStart, bool negative) {
-  auto negOrSelf = [negative](int64 val) { return negative ? ~val + 1 : val; };
+  auto negOrSelf = [negative](uint64 val) { return negative ? ~val + 1 : val; };
 
   theIntVal = static_cast<int64>(HexCharToDigit(GetCharAtWithUpperCheck(curIdx)));
 
@@ -255,8 +255,9 @@ TokenKind MIRLexer::GetIntConst(uint32 valStart, bool negative) {
 
   char c = GetNextCurrentCharWithUpperCheck();
 
-  for (theIntVal = negOrSelf(theIntVal); isdigit(c); c = GetNextCurrentCharWithUpperCheck()) {
-    theIntVal = (theIntVal * radix) + negOrSelf(static_cast<int64>(HexCharToDigit(c)));
+  for (theIntVal = static_cast<int64>(negOrSelf(static_cast<uint64>(theIntVal)));
+      isdigit(c); c = GetNextCurrentCharWithUpperCheck()) {
+    theIntVal = (theIntVal * radix) + static_cast<int64>(negOrSelf(static_cast<uint32>(HexCharToDigit(c))));
   }
 
   if (c == 'u' || c == 'U') {  // skip 'u' or 'U'
