+++ src/mapleall/maple_me/src/lmbc_memlayout.cpp
@@ -31,8 +31,8 @@ void LMBCMemLayout::LayoutStackFrame(void) {
   }
 
   // allocate the local variables
-  uint32 symtabsize = func->GetSymTab()->GetSymbolTableSize();
-  for (uint32 i = 0; i < symtabsize; i++) {
+  size_t symtabsize = func->GetSymTab()->GetSymbolTableSize();
+  for (size_t i = 0; i < symtabsize; i++) {
     MIRSymbol *sym = func->GetSymTab()->GetSymbolFromStIdx(i);
     if (!sym) {
       continue;
@@ -60,12 +60,12 @@ void LMBCMemLayout::LayoutStackFrame(void) {
 
 GlobalMemLayout::GlobalMemLayout(MIRModule *mod, MapleAllocator *mallocator)
     : seg_GPbased(MS_GPbased), sym_alloc_table(mallocator->Adapter()), mirModule(mod) {
-  uint32 symtabsize = GlobalTables::GetGsymTable().GetSymbolTableSize();
+  size_t symtabsize = GlobalTables::GetGsymTable().GetSymbolTableSize();
   sym_alloc_table.resize(symtabsize);
   MIRSymbol *sym = nullptr;
   // allocate the global variables ordered based on alignments
   for (uint32 curalign = 8; curalign != 0; curalign >>= 1) {
-    for (uint32 i = 0; i < symtabsize; i++) {
+    for (size_t i = 0; i < symtabsize; i++) {
       sym = GlobalTables::GetGsymTable().GetSymbolFromStidx(i);
       if (!sym) {
         continue;
@@ -86,7 +86,7 @@ GlobalMemLayout::GlobalMemLayout(MIRModule *mod, MapleAllocator *mallocator)
       seg_GPbased.size += static_cast<int32>(sym->GetType()->GetSize());
     }
   }
-  seg_GPbased.size = maplebe::RoundUp(seg_GPbased.size, GetPrimTypeSize(PTY_ptr));
+  seg_GPbased.size = static_cast<int32>(maplebe::RoundUp(seg_GPbased.size, GetPrimTypeSize(PTY_ptr)));
   mirModule->SetGlobalMemSize(seg_GPbased.size);
 }
